requirements.txt
torch
torchvision
opencv-python
flask
pillow
pillow-heif
numpy
scikit-learn
tqdm
gradio
